src/test/java/com/blank04/universitycms/service/AudienceServiceTest.java
package com.blank04.universitycms.service;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.context.annotation.ComponentScan;
import org.springframework.context.annotation.FilterType;
import org.springframework.test.context.jdbc.Sql;

import java.sql.SQLException;
import java.util.ArrayList;

import static org.junit.jupiter.api.Assertions.*;

@DataJpaTest(includeFilters = {@ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = {AudienceService.class})})
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
@Sql(
        scripts = {"/sql/clear_tables.sql", "/sql/sample_data.sql"},
        executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD
)
class AudienceServiceTest {

    @Autowired
    AudienceService audienceService;

    @Test
    void shouldDeleteById() throws SQLException {
        audienceService.deleteById(1000L);
        audienceService.deleteById(1001L);
        assertEquals(new ArrayList<>(), audienceService.findAll());
    }

    @Test
    void shouldNotDeleteById() {
        Exception thrown = assertThrows(SQLException.class, () -> audienceService.deleteById(1L));
        assertEquals("Audience with id=1 doesn't exist", thrown.getMessage());
    }
}
